+++ src/main/java/com/uaes/esw/gwmc30demo/domain/model/vehicle/Battery.java
@@ -5,5 +5,5 @@ import lombok.Data;
 
 @Data @Builder
 public class Battery {
-    private int soc;
+    private double soc;
 }
